prop some config

+++ common/global_config.go
@@ -5,10 +5,10 @@ import (
 )
 
 type GlobalConfig struct {
-	NumBuckets    uint32
-	BucketDepth   uint32
-	WindowSize    uint64
-	DataSize      uint32 // Number of bytes
+	NumBuckets    int
+	BucketDepth   int
+	WindowSize    int
+	DataSize      int // Number of bytes
 	WriteInterval time.Duration
 	ReadInterval  time.Duration
 	TrustDomains  []*TrustDomainConfig
